+++ core/com.b2international.snowowl.core.rest.tests/src/com/b2international/snowowl/core/rest/auth/AuthorizationTest.java
@@ -15,30 +15,19 @@
  */
 package com.b2international.snowowl.core.rest.auth;
 
-import static com.b2international.snowowl.core.rest.BundleApiAssert.createBundle;
-import static com.b2international.snowowl.core.rest.BundleApiAssert.prepareBundleCreateRequestBody;
-import static com.b2international.snowowl.core.rest.CodeSystemApiAssert.assertCodeSystemCreate;
 import static com.b2international.snowowl.test.commons.ApiTestConstants.CODESYSTEMS_API;
-import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.equalTo;
 
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
 
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
 import com.b2international.commons.json.Json;
 import com.b2international.snowowl.core.codesystem.CodeSystem;
-import com.b2international.snowowl.core.domain.IComponent;
 import com.b2international.snowowl.core.identity.Permission;
 import com.b2international.snowowl.core.internal.ResourceDocument;
-import com.b2international.snowowl.core.request.ResourceRequests;
-import com.b2international.snowowl.snomed.common.SnomedConstants.Concepts;
-import com.b2international.snowowl.snomed.common.SnomedTerminologyComponentConstants;
 import com.b2international.snowowl.test.commons.ApiTestConstants;
-import com.b2international.snowowl.test.commons.Services;
+import com.b2international.snowowl.test.commons.auth.BaseAuthorizationTest;
 import com.b2international.snowowl.test.commons.rest.RestExtensions;
 
 import io.restassured.http.ContentType;
@@ -46,85 +35,8 @@ import io.restassured.http.ContentType;
 /**
  * @since 8.0
  */
-public class AuthorizationTest {
+public class AuthorizationTest extends BaseAuthorizationTest {
 
-	private static final String SNOMEDCT_ID = "SNOMEDCT";
-	private static final String SNOMEDCT_UK_CL = "SNOMEDCT-UK-CL";
-	private static final String SNOMEDCT_UK_CL_MICRO = "SNOMEDCT-UK-CL-MICRO";
-	private static final String SNOMEDCT_UK_DR = "SNOMEDCT-UK-DR";
-	private static final String UK_ALL_BUNDLE_ID = "uk-all";
-	private static final String UK_CLINICAL_BUNDLE_ID = "uk-clinical";
-	
-	private static final Json SNOMED = Json.object(
-		ResourceDocument.Fields.ID, SNOMEDCT_ID,
-		ResourceDocument.Fields.TITLE, "SNOMED CT",
-		ResourceDocument.Fields.URL, SnomedTerminologyComponentConstants.SNOMED_URI_SCT,
-		ResourceDocument.Fields.TOOLING_ID, SnomedTerminologyComponentConstants.TOOLING_ID,
-		ResourceDocument.Fields.BUNDLE_ID, IComponent.ROOT_ID
-	);
-	
-	@BeforeClass
-	public static void setup() {
-		// create test Bundles
-		createBundle(prepareBundleCreateRequestBody(UK_ALL_BUNDLE_ID));
-		createBundle(prepareBundleCreateRequestBody(UK_CLINICAL_BUNDLE_ID, UK_ALL_BUNDLE_ID));
-		
-		// create test CodeSystems, one in the root, one inside the bundle
-		assertCodeSystemCreate(SNOMED)
-			.statusCode(201)
-			.header("Location", containsString("/codesystems/SNOMEDCT"));
-		
-		assertCodeSystemCreate(
-			Json.object(
-				ResourceDocument.Fields.ID, SNOMEDCT_UK_CL,
-				ResourceDocument.Fields.TITLE, "SNOMED CT UK Clinical",
-				ResourceDocument.Fields.URL, SnomedTerminologyComponentConstants.SNOMED_URI_SCT + "/" + Concepts.UK_CLINICAL_EXTENSION_MODULE,
-				ResourceDocument.Fields.TOOLING_ID, SnomedTerminologyComponentConstants.TOOLING_ID,
-				ResourceDocument.Fields.BUNDLE_ID, UK_CLINICAL_BUNDLE_ID
-			))
-			.statusCode(201)
-			.header("Location", containsString("/codesystems/" + SNOMEDCT_UK_CL));
-		
-		assertCodeSystemCreate(
-			Json.object(
-				ResourceDocument.Fields.ID, SNOMEDCT_UK_CL_MICRO,
-				ResourceDocument.Fields.TITLE, "SNOMED CT UK Clinical (Micro Releases)",
-				ResourceDocument.Fields.URL, SnomedTerminologyComponentConstants.SNOMED_URI_DEV + "/" + Concepts.UK_CLINICAL_EXTENSION_MODULE,
-				ResourceDocument.Fields.TOOLING_ID, SnomedTerminologyComponentConstants.TOOLING_ID,
-				ResourceDocument.Fields.BUNDLE_ID, UK_CLINICAL_BUNDLE_ID
-			))
-			.statusCode(201)
-			.header("Location", containsString("/codesystems/" + SNOMEDCT_UK_CL));
-		
-		assertCodeSystemCreate(
-			Json.object(
-				ResourceDocument.Fields.ID, SNOMEDCT_UK_DR,
-				ResourceDocument.Fields.TITLE, "SNOMED CT UK Drug",
-				ResourceDocument.Fields.URL, SnomedTerminologyComponentConstants.SNOMED_URI_SCT + "/" + Concepts.UK_DRUG_EXTENSION_MODULE,
-				ResourceDocument.Fields.TOOLING_ID, SnomedTerminologyComponentConstants.TOOLING_ID,
-				ResourceDocument.Fields.BUNDLE_ID, UK_ALL_BUNDLE_ID
-			))
-			.statusCode(201)
-			.header("Location", containsString("/codesystems/" + SNOMEDCT_UK_DR));
-	}
-	
-	@AfterClass
-	public static void after() {
-		ResourceRequests
-			.prepareSearch()
-			.all()
-			.buildAsync()
-			.execute(Services.bus())
-			.getSync(1, TimeUnit.MINUTES)
-			.forEach(resource -> {
-				ResourceRequests
-				.prepareDelete(resource.getId())
-				.build(RestExtensions.USER, "Delete " + resource.getId())
-				.execute(Services.bus())
-				.getSync(1, TimeUnit.MINUTES); 
-			});
-	}
-	
 	@Test
 	public void emptyPermissionsNoAccess() throws Exception {
 		String token = RestExtensions.generateToken();
