+++ tests/com.b2international.snowowl.test.commons/META-INF/MANIFEST.MF
@@ -26,6 +26,7 @@ Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-ActivationPolicy: lazy
 Import-Package: org.slf4j;version="1.7.25"
 Export-Package: com.b2international.snowowl.test.commons,
+ com.b2international.snowowl.test.commons.auth,
  com.b2international.snowowl.test.commons.codesystem,
  com.b2international.snowowl.test.commons.inject,
  com.b2international.snowowl.test.commons.rest,
